Code reorganization
Changing the coding style to something more pythonic (avoid lowerCamelCase for functions and method names, for example).
Trying to move some github clients functionalities out of the Jucybot class, and moving them in a GithubClient mixin, so that we can use the same methods when dealing either with the jucybot account and with the logged user account.